split qml into its own installable package
