+++ qstack/qstack-qml/README.md
@@ -1,0 +1,52 @@
+[![Lifecycle:
+experimental](https://img.shields.io/badge/lifecycle-experimental-orange.svg)](https://lifecycle.r-lib.org/articles/stages.html#experimental)
+==============================================
+Q-stack (QML)
+==============================================
+<p align="center"><img alt="qstack logo" align="center" src="./images/logo.png" width=66%></p>
+
+
+## Contents
+* [Contents](#Contents-)
+* [About](#about-)
+* [Install](#install-)
+* [Examples](#examples-)
+* [References](#references-)
+* [Acknowledgements](#acknowledgements-)
+
+## About [↑](#contents)
+
+Q-stack is a stack of codes for dedicated pre- and post-processing tasks for Quantum Machine Learning (QML). It is a work in progress. Stay tuned for updates!
+
+This sub-part of Q-stack contains a few geometry-only descriptors, which are the following ones:
+
+- $B^2R^2$ (https://github.com/lcmd-epfl/b2r2-reaction-rep)
+- $\mathrm{SLATM}_d$
+
+      
+## Install [↑](#contents)
+
+The installation of this sub-part can be done executing the following command:
+
+```
+python -m pip install git+https://github.com/lcmd-epfl/Q-stack.git#subdirectory=qstack/qstack-qml
+```
+## Examples [↑](#contents)
+Q-stack comes with several example codes that illustrate some of its key capabilities. Examples for firs sub-part will follow shortly.
+
+
+## References [↑](#contents)
+
+* P. van Gerwen, A. Fabrizio, M. Wodrich, and C. Corminboeuf,
+  “Physics-based representations for machine learning properties of chemical reactions”,
+  Mach. Learn.: Sci. Technol. **3**, 045005 (2022)
+  [![DOI](https://img.shields.io/badge/DOI-10.1088/2632--2153/ac8f1a-blue)](https://doi.org/10.1088/2632-2153/ac8f1a)
+
+
+## Acknowledgements [↑](#contents)
+The authors of Q-stack acknowledge the National Centre of Competence in Research (NCCR)
+"Materials' Revolution: Computational Design and Discovery of Novel Materials (MARVEL)" 
+of the Swiss National Science Foundation (SNSF, grant number 182892)
+and the European Research Council (ERC, grant agreement no 817977).
+
+![acknowledgements logos](./images/ackw.png)
